requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
idna==3.3
PyJWT==2.3.0
pymongo==3.12.3
pyotp==2.6.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.2.4
twilio==7.7.0
urllib3==1.26.8
